requirements.txt
python-telegram-bot==20.7
requests
python-dotenv
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.2
fastapi==0.103.2
uvicorn==0.23.2  
python-jose
